requirements.txt
fastapi==0.104.1
sqlmodel==0.0.14
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
databases[sqlite]==0.8.0
databases[postgresql]==0.8.0
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.12.1
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
pre-commit==3.5.0

# Utilities
pydantic[email]==2.5.0
email-validator==2.1.0
phonenumbers==8.13.26
